simulacao-consignado/src/main/java/com/example/simulacaoconsignado/Entites/Cliente.java
package com.example.simulacaoconsignado.Entites;

public class Cliente {

        private String cpf;
        private boolean correntista;
        private String segmento;

    public String getCpf() {
        return cpf;
    }

    public void setCpf(String cpf) {
        this.cpf = cpf;
    }

    public boolean isCorrentista() {
        return correntista;
    }

    public void setCorrentista(boolean correntista) {
        this.correntista = correntista;
    }

    public String getSegmento() {
        return segmento;
    }

    public void setSegmento(String segmento) {
        this.segmento = segmento;
    }

    public String getConvenio() {
        return convenio;
    }

    public void setConvenio(String convenio) {
        this.convenio = convenio;
    }

    private String convenio;
}
